+++ client/styles/Home.css
@@ -1,7 +1,8 @@
-.lobby-container {
+.home-container {
   width: 100vw;
   height: 100vh;
   display: flex;
+  flex-direction:column;
   justify-content: center;
   align-items: center;
   align-content: center;
