+++ public/index.html
@@ -1,10 +1,11 @@
 <html>
   <head>
-    <script defer src='/main.js'></script>
+    <script defer src="/main.js"></script>
+    <link rel="stylesheet" type="text/css" href="style.css" />
   </head>
   <body>
     <div>
-      <div id='root'></div>
+      <div id="root"></div>
     </div>
   </body>
 </html>
